refactor(deprecate): move deprecated to deprecation
- move the `deprecated` front-door dispatcher (and its class-path helpers `_packing_class_source` and `_PackingClassArgs`) from `routine.py` into `deprecation.py`, so `from deprecate.deprecation import deprecated` resolves natively instead of via a shim
- `routine.py` now holds only the callable-specific `deprecated_callable` plus the packing descriptors; `deprecation.py` imports `deprecated_callable` from it (one-way, no cycle)
- repoint `__init__`, the api-reference autodoc path, and the structure-tree docs

---
